DIGITAL CONSORTIUM
Blockchain-based crowdfunding invites you to cooperate, providing support for ambitious projects with capital backed by 3 trillion rubles in securities. This is a unique opportunity to become part of an innovative ecosystem and influence the future of blockchain technologies.

By participating in crowdfunding, you get not only financial benefits, but also the opportunity to actively participate in project management. You are provided with management tokens that give you the right to vote in making key decisions regarding the development and strategy of projects.

Moreover, each participant receives an NFT share in promising projects, opening up access to unique assets and opportunities. This is not just an investment, it is a participation in creating a new world where technology and community come together to achieve common goals.

And finally, your contribution will not go unnoticed. Participation in crowdfunding on the blockchain PLATFORM is a chance to gain worldwide fame, become part of the global community of innovators and contribute to the development of future technologies. Join us and become a part of the story!